+++ packages/engine/Source/Core/Matrix4.js
@@ -9,6 +9,7 @@ import DeveloperError from "./DeveloperError.js";
 import CesiumMath from "./Math.js";
 import Matrix3 from "./Matrix3.js";
 import RuntimeError from "./RuntimeError.js";
+import deprecationWarning from "./deprecationWarning.js";
 
 /** @import Quaternion from "./Quaternion.js"; */
 /** @import TranslationRotationScale from "./TranslationRotationScale.js"; */
@@ -709,11 +710,17 @@ class Matrix4 {
   /**
    * Computes a Matrix4 instance from a Camera.
    *
+   * @deprecated This method has been deprecated and will be removed in Cesium 1.151. Use {@link Camera#viewMatrix} instead.
+   *
    * @param {CameraLike} camera The camera to use.
    * @param {Matrix4} [result] The object in which the result will be stored, if undefined a new instance will be created.
    * @returns {Matrix4} The modified result parameter, or a new Matrix4 instance if one was not provided.
    */
   static fromCamera(camera, result) {
+    deprecationWarning(
+      "Matrix4.fromCamera",
+      "This method has been deprecated and will be removed in Cesium 1.151. Use Camera.prototype.viewMatrix instead.",
+    );
     //>>includeStart('debug', pragmas.debug);
     Check.typeOf.object("camera", camera);
     //>>includeEnd('debug');